ic_stable_structures::storable

Trait Storable

pub trait Storable {
    const BOUND: Bound;

    // Required methods
    fn to_bytes(&self) -> Cow<'_, [u8]>;
    fn from_bytes(bytes: Cow<'_, [u8]>) -> Self;

    // Provided method
    fn to_bytes_checked(&self) -> Cow<'_, [u8]> { ... }
}

A trait with convenience methods for storing an element into a stable structure.

Required Associated Constants

const BOUND: Bound

The size bounds of the type.

Required Methods

fn to_bytes(&self) -> Cow<'_, [u8]>

Converts an element into bytes.

NOTE: Cow is used here to avoid unnecessary cloning.

fn from_bytes(bytes: Cow<'_, [u8]>) -> Self

Converts bytes into an element.

Provided Methods

fn to_bytes_checked(&self) -> Cow<'_, [u8]>

Like to_bytes, but includes additional checks to ensure the element’s serialized bytes are within the element’s bounds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.