ic_stable_structures

Struct RestrictedMemory

pub struct RestrictedMemory<M: Memory> { /* private fields */ }

RestrictedMemory creates a limited view of another memory. This allows one to divide the main memory into non-intersecting ranges and use different layouts in each region.

Implementations

impl<M: Memory> RestrictedMemory<M>

pub fn new(memory: M, page_range: Range<u64>) -> Self