ic_stable_structures::vec

Struct Vec

pub struct Vec<T: Storable, M: Memory>(/* private fields */);

An implementation of growable arrays in stable memory.

Implementations

impl<T: Storable, M: Memory> Vec<T, M>

pub fn new(memory: M) -> Result<Self, GrowFailed>

Creates a new empty vector in the specified memory, overwriting any data structures the memory might have contained previously.

Complexity: O(1)

pub fn init(memory: M) -> Result<Self, InitError>

Initializes a vector in the specified memory.

Complexity: O(1)

PRECONDITION: the memory is either empty or contains a valid stable vector.

pub fn into_memory(self) -> M

Returns the underlying memory instance.

pub fn is_empty(&self) -> bool

Returns true if the vector is empty.

Complexity: O(1)

pub fn len(&self) -> u64

Returns the number of items in the vector.

Complexity: O(1)

pub fn set(&self, index: u64, item: &T)

Sets the item at the specified index to the specified value.

Complexity: O(max_size(T))

PRECONDITION: index < self.len()

pub fn get(&self, index: u64) -> Option<T>

Returns the item at the specified index.

Complexity: O(max_size(T))

pub fn push(&self, item: &T) -> Result<(), GrowFailed>

Adds a new item at the end of the vector.

Complexity: O(max_size(T))

pub fn pop(&self) -> Option<T>

Removes the item at the end of the vector.

Complexity: O(max_size(T))

pub fn iter(&self) -> impl DoubleEndedIterator<Item = T> + '_